Revert "Better support for nesting of parallel regions:"
This commit was fundamentally broken, because the concept of
"main thread" is much more involved than originally thought.

To work correctly, the heirarchical structure of parallel regions
must be either discovered or specified by the user.

This commit broke ex2, for example, because thread 2 should be the
main thread for TASK_G_0.

Also, beware that lambdas can vary between threads (local variables),
so it is not clear what to do when multiple threads can start a
parallel region.

This reverts commit 410ec8f1c792fff52e7fdef021200a24cd5710a3.