Change how default scheduling can be used with STS instances.
Users can no longer create instances with default scheduling, but
that is okay because it is available through the default instance.

This makes it much clearer what instances require calling "nextStep"
and "wait" (all but the default instance). Also, this means that new
instances use a single thread for unassigned loops. No automatic
splitting is done except by the default instance.