Allow users to set any STS instance to use the default schedule
This is useful for reverting to a simple schedule when needed (say
to create a base case for comparison, for testing, or for cases
where thread assignments aren't helpful).

Also, boolean added to indicate if STS instance is "active." This
allows for more asserts, better error checking, and STS can now
tolerate multiple calls to nextStep and wait.