Manage subtasks with a call stack instead of a next-subtask index
Replace the per-thread index for the next subtask with a per-thread
call stack of running subtasks. This change allows us to simplify
the logic for iterating through and nesting subtasks.

Also, simplify how tasks are skipped, by just setting the "isDone"
flag rather than running an empty functor. Maintaining the old
behavior would have complicated the refactoring, because we have to
check for skipping as a special case, and it is no longer necessary
with the "isDone" flag available.