Change scheduling algorithm to be simpler and clearer
It is simpler in that pause targets can now only be later tasks,
greatly simplifying the logic, and tasks are always executed in
linear order outside of pausing.

It is clearer in that there is now a strong guarantee about
checkpointing. Subtasks are never re-entered before a cp is reached,
which should help application code to be simpler and avoid deadlocks.